+++ src/main/java/racingcar/dto/GamePlayRequestDto.java
@@ -3,9 +3,9 @@ package racingcar.dto;
 public class GamePlayRequestDto {
 
     private final String names;
-    private final String count;
+    private final int count;
 
-    public GamePlayRequestDto(final String names, final String count) {
+    public GamePlayRequestDto(final String names, final int count) {
         this.names = names;
         this.count = count;
     }
@@ -14,7 +14,7 @@ public class GamePlayRequestDto {
         return names;
     }
 
-    public String getCount() {
+    public int getCount() {
         return count;
     }
 }
